fix: separate hash
